What it attempts to do: - calculate the real totals and update the AttendeeResults table.
Pushing broken code onto this branch, to get help with it at the elixir meetup. I'm trying to collect a list of changesets and make a single call to Multi.update_all to update their rows in the database. It gets this error:
```
function Ecto.Multi.update_all/3 is undefined or private. Did you mean one of:
      * update_all/4
      * update_all/5
```

Where I plan to clean it up (and don't want much advice on yet):
- DRY the code
- move queries into a composable place (the model? the schema?)
 - move actions to other controllers, especially the "wizard" ones of:
   - picking a night (currently GameNightController:take_attendance)
   - naming the attendees (currently GameNightController:current_attendance)
   - counting chips (currently AttendeeResultController:add_chips and AttendeeResultController:subtract_chips)
   - displaying results (currently GameNightController:cash_out and GameNightController:cash_out_player)
   - sending results? (GameNightController:send_results)
- generally put stuff where it belongs, in an elixir-idiomatic way